Memory leak when upgrading from tcp frontend to http backend
The command:
  # echo "show pools" | socat - /tmp/haproxy_admin.sock
shows that a pool, in my case the "email_alert", is constantly growing.

Issue description:
The tcp frontend forwards to an HTTP backend. The stream_set_backend
function is called (proxy.c line 1404) and then the conn_upgrade_mux_fe
(include/proto/connection.h, line 1158) to upgrade the client mux.
A new mux is build and the old one is destroyed with a mux_pt_destroy
(mux_pt.c, line 27) function call. Inside the mux_pt_destroy function the
the ctx->wait_event.tasklet is never released  because of the
"ctx->conn->ctx == ctx" check which fails.